[BE] [fix] orderDto 수정

See merge request s11-fintech-finance-sub1/S11P21E202!201

+++ backend/givutake/src/main/java/com/accepted/givutake/gift/controller/GiftController.java
@@ -163,7 +163,7 @@ public class GiftController {
     @GetMapping("/review/order/{orderIdx}")
     public ResponseEntity<ResponseDto> getUserReviewOrder(
             @AuthenticationPrincipal UserDetails userDetails,
-            @PathVariable int orderIdx){
+            @PathVariable long orderIdx){
         boolean data = giftService.IsWriteGiftReview(userDetails.getUsername(), orderIdx);
         ResponseDto responseDto = ResponseDto.builder()
                 .data(data)
